Correct the name of the imported pip package for Fedora 39
Right now the version of pip that ships with Fedora 39 _does not_
insist on normalizing the package name.  I think this may be new
behavior as of pip version 23.2, which was the first version to
include the changes in pypa/pip#12044.  (This PR was a fix to address
pypa/pip#12038.)

See also:
https://pip.pypa.io/en/stable/news/#v23-2